pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = ["hatchling", "hatch-vcs"]

[project]
dependencies = ["rich_argparse", "pybids", "nibabel"]
dynamic = ["version"]
license = {text = "MIT"}
name = "cat12"
readme = "README.md"
requires-python = ">=3.9"

[project.optional-dependencies]
docs = [
    "myst-parser",
    "sphinx",
    "sphinx-argparse",
    "sphinx-copybutton",
    "pydata-sphinx-theme",
    "sphinx-togglebutton"
]

[tool.codespell]
builtin = "clear,rare"

[tool.hatch.build.hooks.vcs]
version-file = "src/cat12/_version.py"

[tool.hatch.build.targets.wheel]
packages = ["src/cat12"]

[tool.hatch.version]
source = "vcs"

[tool.ruff]
include = ["pyproject.toml", "src/**/*.py", "scripts/**/*.py"]
indent-width = 4
line-length = 79
target-version = "py39"

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = "dynamic"
indent-style = "space"
line-ending = "auto"
quote-style = "double"
skip-magic-trailing-comma = false

[tool.ruff.lint]
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
fixable = ["ALL"]
ignore = []
select = ["E", "F", "W", "C90", "I", "D", "B", "UP", "N", "ARG", "PTH", "FLY", "RUF"]
unfixable = []

[tool.ruff.lint.mccabe]
max-complexity = 20

[tool.ruff.lint.pydocstyle]
convention = "numpy"
